Add empty document with grid and baseline
I chose to use a US Letter page to simplify the grid system using points,
which is much more convenient for alignment with the baseline grid.

Because 1 inch = 72 points = 25.4mm, it is easy to manipulate multiples
of 12pt (the typical printer margins) with inches, while it is really
inconvenient in mm (12pt ~ 4.233333333mm). An interesting conversion when
working in mm is 9pt = 3.175mm. It does not help much with the fact that
the A4 format corresponds to 597.276 by 841.892 pt...