Adjust styles of headings and layout of last article
The last article was not properly aligned with the baseline
in the corner: it was on a secondary line. I also extended
the ruler to hover over the whole article.

I adjusted the font size of headings, and replaced expansion
of the letter forms with increase of the tracking space
between letters. I also increased the dimensions of the text
boxes for headings to allow their full coverage of the width of
the column without avoiding their disappance because of overflow.

I have changed the typeface used for the title "In Memoriam",
which I changed to all caps to match the solemnity of the topic.