Export to PDF and compress using Adobe online tool
The online tool looks now available for free on a permanent basis.
It was previously part of the Adobe Acrobat Pro DC cloud subscription,
and made available for free for a limited period of time.
The tool is now advertised on its own, with a reference to Adobe Pro DC
for more control on the compression.

  Adobe Acrobat: Compress PDF
  https://www.adobe.com/acrobat/online/compress-pdf.html